+++ include/H2DE/H2DE_camera.h
@@ -1,7 +1,7 @@
 #ifndef H2DE_CAMERA_H
 #define H2DE_CAMERA_H
 
-#include "H2DE_engine.h"
+#include <H2DE_engine.h>
 class H2DE_Engine;
 
 class H2DE_Camera {
@@ -16,7 +16,7 @@ public:
 
     friend H2DE_LevelPos H2DE_GetCameraPos(H2DE_Camera* camera); 
     friend H2DE_LevelSize H2DE_GetCameraSize(H2DE_Camera* camera);
-    friend bool H2DE_CameraContains(H2DE_Camera* camera, H2DE_LevelHitbox hitbox);
+    friend bool H2DE_CameraContains(H2DE_Camera* camera, H2DE_LevelRect hitbox);
 
     friend void H2DE_SetCameraPos(H2DE_Camera* camera, H2DE_LevelPos pos);
     friend void H2DE_SetCameraSize(H2DE_Camera* camera, float blockOnWidth);