+++ include/H2DE/H2DE_game_data.h
@@ -1,0 +1,9 @@
+#ifndef H2DE_GAME_DATA_H
+#define H2DE_GAME_DATA_H
+
+struct H2DE_GameData {
+    float gravity = 0.01f;
+    float maxGravity = 0.35f;
+};
+
+#endif